NASAEARTH-1 fix dynamic dev server port in playwright ui
In Playwright ui the project dependencies are not run.
